refactor(acceptMethod-util): simplify and rename methodRouter  to acceptMethod
